+++ lib/src/entry.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import 'package:meta/meta.dart';
+
 import 'constants.dart';
 import 'header.dart';
 
@@ -8,10 +10,21 @@ import 'header.dart';
 /// Usually, tar entries are read from a stream, and they're bound to the stream
 /// from which they've been read. This means that they can only be read once,
 /// and that only one [TarEntry] is active at a time.
-class TarEntry extends Stream<List<int>> {
+@sealed
+class TarEntry {
   /// The parsed [TarHeader] of this tar entry.
   final TarHeader header;
-  final Stream<List<int>> _dataStream;
+
+  /// The content stream of the active tar entry.
+  ///
+  /// This is a single-subscription stream backed by the original stream used to
+  /// create a tar reader.
+  /// When listening on [contents], the stream needs to be fully drained before
+  /// the next call to [StreamIterator.next]. It's acceptable to not listen to
+  /// [contents] at all before calling [StreamIterator.next] again. In that
+  /// case, this library will take care of draining the stream to get to the
+  /// next entry.
+  final Stream<List<int>> contents;
 
   /// The name of this entry, as indicated in the header or a previous pax
   /// entry.
@@ -26,17 +39,10 @@ class TarEntry extends Stream<List<int>> {
   /// Time of the last modification of this file, as indicated in the [header].
   DateTime get modified => header.modified;
 
-  TarEntry(this.header, this._dataStream);
+  TarEntry(this.header, this.contents);
 
   /// Creates an in-memory tar entry from the [header] and the [data] to store.
   factory TarEntry.data(TarHeader header, List<int> data) {
     return TarEntry(header, Stream.value(data));
-  }
-
-  @override
-  StreamSubscription<List<int>> listen(void Function(List<int> event)? onData,
-      {Function? onError, void Function()? onDone, bool? cancelOnError}) {
-    return _dataStream.listen(onData,
-        onError: onError, onDone: onDone, cancelOnError: cancelOnError);
   }
 }